requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
Authlib==1.3.1
Babel==2.9.1
bcrypt==3.1.7 
black==24.1.1
click==8.1.7
cryptography==3.4.8
email_validator==2.1.2
fastapi==0.115.0
celery==5.3.6
redis==5.0.1
flake8==7.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-oauth==0.15.1
idna==3.6
Jinja2==3.0.3
makefun==1.15.6
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
phonenumbers==8.13.43
Pillow==9.0.1
platformdirs==4.2.0
psycopg2==2.9.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==21.0.0
pyparsing==3.1.1
PyPDF2==1.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
sentry-sdk==2.16.0
SQLAlchemy==2.0.25
sqlmodel==0.0.22
starlette==0.38.6
typing_extensions==4.9.0
urllib3==2.2.3
uvicorn==0.27.1
aiofiles==24.1.0
asyncpg>=0.27.0
pyfcm==2.0.8
pusher==3.3.3
boto3==1.28.33
python-multipart==0.0.20
aioredis==2.0.1
